+++ desktop/appcontainer/ToastSource.js
@@ -9,8 +9,9 @@ import {div} from '@xh/hoist/cmp/layout';
 import {hoistCmp, HoistModel, useLocalModel, uses} from '@xh/hoist/core';
 import {Position, Toaster} from '@xh/hoist/kit/blueprint';
 import {withDefault} from '@xh/hoist/utils/js';
-import {isElement} from 'lodash';
+import {isElement, map} from 'lodash';
 import './Toast.scss';
+import {wait} from '../../promise';
 
 /**
  *  Support for showing Toasts in a application. This component does not render any content
@@ -35,8 +36,8 @@ class LocalModel extends HoistModel {
     constructor(toastSourceModel) {
         super();
         this.addReaction({
-            track: () => toastSourceModel.toastModels,
-            run: this.displayPendingToasts
+            track: () => [toastSourceModel.toastModels, map(toastSourceModel.toastModels, 'isOpen')],
+            run: ([models]) => this.displayPendingToasts(models)
         });
     }
 
@@ -45,18 +46,24 @@ class LocalModel extends HoistModel {
     //------------------------
     displayPendingToasts(models) {
         models.forEach(model => {
-            let {wasShown, isOpen, icon, position, containerRef, ...rest} = model;
-            if (wasShown || !isOpen) return;
-
+            let {bpId, isOpen, icon, position, containerRef, ...rest} = model;
             position = position || Position.BOTTOM_RIGHT;
-            this.getToaster(position, containerRef).show({
-                className: 'xh-toast',
-                icon: div({className: 'xh-toast__icon', item: icon}),
-                onDismiss: () => model.dismiss(),
-                ...rest
-            });
 
-            model.wasShown = true;
+            // 1) If toast is visible and sent to bp, or already obsolete -- nothing to do
+            if ((!!bpId) === isOpen) return;
+
+            // 2) ...otherwise this toast needs to be shown or hidden with bp api
+            let toaster = this.getToaster(position, containerRef);
+            if (!bpId) {
+                model.bpId = toaster.show({
+                    className: 'xh-toast',
+                    icon: div({className: 'xh-toast__icon', item: icon}),
+                    onDismiss: () => wait(0).then(() => model.dismiss()),
+                    ...rest
+                });
+            } else {
+                toaster.dismiss(bpId);
+            }
         });
     }